Reorganize main loop into app.py
Using dependency injection to avoid circular import.
